+++ src/components/nvi/export-buttons/ExportNviPublicationPointsButton.tsx
